+++ encoding/protobuf/error.go
@@ -95,7 +95,7 @@ func GetErrorDetails(err error) []interface{} {
 		return nil
 	}
 	var target *pberror
-	if errors.As(err, &target) {
+	if errors.As(err, &target) && len(target.details) > 0 {
 		results := make([]interface{}, 0, len(target.details))
 		for _, any := range target.details {
 			detail := &types.DynamicAny{}
@@ -154,7 +154,7 @@ func createStatusWithDetail(pberr *pberror, encoding transport.Encoding, codec *
 	st := status.New(grpcerrorcodes.YARPCCodeToGRPCCode[pberr.code], pberr.message)
 	// Here we check that status.New has returned a valid error.
 	if st.Err() == nil {
-		return nil, errors.New("no status error for error with code OK")
+		return nil, errors.New(fmt.Sprintf("no status error for error with code %d", pberr.code))
 	}
 	pst := st.Proto()
 	pst.Details = pberr.details